MA-9266 [Android_MX6Q_SD] Can not switch to audio playback low power mode.100%
The file cpufreq/interactive/input_boost is changed to cpufreq/interactive/boost in new system.
The do_changecpugov function cannot find the file with old file name and return.
When cpu prepare to enter conservative mode(gov=conservative), as system cannot find the file
and return, the function write(mFd, gov, strlen(gov)) is not excuted. So system cannot enter
conservative mode.

When system work in interactive mode, the core frequency will jump to maximum frequency frequently
because of the workload variation. The busfreq can drop down the bus frequency to 50MHz after 3second
only when the core frequency is not maximum frequency and other requirements are satisfied.

When system work in conservative mode, the core frequency will change step by step, not directly to
maximum frequency. So it's easy to satisfy the requirements for busfreq to drop down the bus frequency.

Change-Id: Id0299b663f43c7037ce8f44a9943aee8f8edfe2f
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/1604
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>