MA-9397[Android] Correct the coverity issue 25560
The boost method is not available if the cpufreq is not in interactive mode.
It's no need to disable boost when leave interactive mode.

Enable boostpulse function when enter interactive mode.
It will boost the cpu frequency to maximum and retain it for boostpulse_duration time.
then cpufreq will adjust the frequency according to workload.

Change-Id: I31c4126140f23b5ef2e2dd504b919dcaa7c6377e
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/1639
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>